refactor and deprecate old equiformerv2 (#812)
* refactor and deprecate old equiformerv2

* remove equiv2_backbone_and_heads

* lint fixes

* remove backbone and heads model

* fix merge

* split up tests

* update

* add in missing file
